Use Packrat
It's going to be important to keep our dependencies in check, given the
sheer number we'll be invoking in the process of training the ensemble.
This should ensure that our results continue to be reproducible with the
same code years down the line, even if the underlying ML packages
undergo substantial user-facing changes (or even backend changes in
terms of where and when RNGs are called).