Generalize function to merge NMC and dyadic data
We need to do this not just when constructing the training and test sets
but also when computing predictions for all possible dyads, so the
function has been split into its own file.  The function also computes
the logged capability ratio after merging, so as to reduce on some code
reuse.

NB - Confirmed that the output of `05-mid.r` is identical (via
`identical()`) when using the old and new merging functions