Fix storage of replication results
Previously was saving the raw data and the train() output, usually (in
case of glm's) with trimmed output.  Problem was that extracting the
individual log-likelihoods became difficult, particularly if there were
any missing observations in the raw data that ended up being dropped in
model fitting.

Now explicitly stores individual log-likelihoods, as well as the data
used in fitting, the cross-validation results (not the full train()
output), and some other things that might be useful for summarizing or
examining the replication results.