Use all data for training
The original motivation for the training-test split was to get an
unbiased estimate of out-of-sample predictive power.  In a data-poor
environment such as we have, that seems to put us on the wrong end of
the bias-variance tradeoff.  Instead, will use the Tibshirani &
Tibshirani bias correction to obtain an approximately unbiased estimate
of the final super learner's out-of-sample log loss.

n.b.: the manuscript still refers to the training-test split and will
need to be fixed accordingly