Rework model training
There seems to be a memory leak somewhere that was causing the cluster
run over MPI to fail, so the ensemble training has been split into
separate scripts.  They can be run in either order, since we don't need
the full-sample fits to calculate the optimal weights, and vice versa.

In addition, the weight fitting no longer runs over MPI.  Instead we
need to call

    Rscript 15-train-weights.r i

for each i from 1 to 10 (or whatever the number of imputations is).