Fix(?) weird behavior in undirected calculations
Due to what I can only infer is a weird dplyr bug, sometimes the
calculation of undirected DOE scores works, other times it doesn't.
When it doesn't, it looks like it's because one of the columns has been
duplicated in the output.  But the bug isn't completely reproducible ---
running the same block of code twice often fixes it.  I suspect it's a
memory-related bug because sometimes it segfaults too.

The stopifnot() requiring rows to sum to unity seems to catch the bug
when it occurs, but it's worth taking extra care with
32-collect-predict.r going forward.