Make ternary plots of out-of-fold predictions
This is as close as we can get to out-of-sample predictions now that we
don't have a test set.  This requires some changes to the
infrastructure, since now in the weight training we need to save *all*
of the out-of-fold predictions, not just those for the observed
outcomes.

This code is a bit hackish and needs to be cleaned up more later.