Use rpart instead of rpart2
Running CART via rpart2 resulted in hard-to-debug errors when
calculating out-of-fold predictions in the variable importance study.
These arose because of how caret's methods for rpart2 try to approximate
the complexity parameter as a function of maxdepth.  Since rpart tunes
directly over the complexity parameter, using it sidesteps these issues.